게시판 글 등록 , 상세 조회 까지 가능
