+++ drinking01/src/main/java/com/first/controller/BoardController.java
@@ -9,6 +9,7 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.ModelAttribute;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.servlet.mvc.support.RedirectAttributes;
 
 import com.first.domain.BoardVO;
@@ -49,6 +50,14 @@ public class BoardController {
 		rttr.addFlashAttribute("result", boardVO.getBno());//일회성
 		
 		return "redirect:/board/list";
+		
+	}
+	
+	@GetMapping("/get")
+	public void getBoard(@RequestParam("bno") Long bno , Model model)
+	{
+		BoardVO board = service.getBoard(bno);
+		model.addAttribute("board" , board);
 		
 	}
