+++ drinking01/src/main/java/com/first/mapper/BoardMapper.java
@@ -9,5 +9,7 @@ public interface BoardMapper {
 	public List<BoardVO> getList();
 	
 	public void insert(BoardVO boardVO);
+	
+	public BoardVO getBoard(Long bno);
 
 }
