--- drinking01/src/main/java/com/springmvc/model/MemberVO.java
@@ -1,90 +1,0 @@
-package com.springmvc.model;
-
-public class MemberVO {
-	
-	//회원 id
-	private String memberId;
-	
-	//회원 비밀번호
-	private String memberPw;
-	
-	//회원 이름
-	private String memberName;
-	
-	//회원 이메일
-	private String memberEmail;
-	
-	//회원 전화번호
-	private String memberPhone;
-	
-	//회원 상세주소
-	private String memberAddress;
-	
-	//등록일자
-	private int regDate;
-
-	public String getMemberId() {
-		return memberId;
-	}
-
-	public void setMemberId(String memberId) {
-		this.memberId = memberId;
-	}
-
-	public String getMemberPw() {
-		return memberPw;
-	}
-
-	public void setMemberPw(String memberPw) {
-		this.memberPw = memberPw;
-	}
-
-	public String getMemberName() {
-		return memberName;
-	}
-
-	public void setMemberName(String memberName) {
-		this.memberName = memberName;
-	}
-
-	public String getMemberEmail() {
-		return memberEmail;
-	}
-
-	public void setMemberEmail(String memberEmail) {
-		this.memberEmail = memberEmail;
-	}
-
-	public String getMemberPhone() {
-		return memberPhone;
-	}
-
-	public void setMemberPhone(String memberPhone) {
-		this.memberPhone = memberPhone;
-	}
-
-	public String getMemberAddress() {
-		return memberAddress;
-	}
-
-	public void setMemberAddress(String memberAddress) {
-		this.memberAddress = memberAddress;
-	}
-
-	public int getRegDate() {
-		return regDate;
-	}
-
-	public void setRegDate(int regDate) {
-		this.regDate = regDate;
-	}
-
-	@Override
-	public String toString() {
-		return "MemberVO [memberId=" + memberId + ", memberPw=" + memberPw + ", memberName=" + memberName
-				+ ", memberEmail=" + memberEmail + ", memberPhone=" + memberPhone + ", memberAddress=" + memberAddress
-				+ ", regDate=" + regDate + "]";
-	}
-	
-	
-}
