+++ drinking01/src/main/resources/com/first/mapper/BoardMapper.xml
@@ -7,13 +7,17 @@
 
 <select id="getList" resultType="com.first.domain.BoardVO">
 <![CDATA[
-select * from tbl_board where bno > 0
+select * from table_board where bno > 0
 ]]>
 </select>
 
 <select id = "insert" >
-insert into tbl_board(bno , title , content , writer )
+insert into table_board(bno , title , content , writer )
     values (seq_board.nextval , #{title} , #{content} , #{writer})
+</select>
+
+<select id = "getBoard" resultType = "com.first.domain.BoardVO">
+	select * from table_board where bno = #{bno}
 </select>
 
 </mapper>
