+++ drinking01/src/main/webapp/WEB-INF/spring/root-context.xml
@@ -19,8 +19,8 @@
 			value="net.sf.log4jdbc.sql.jdbcapi.DriverSpy"></property>
 		<property name="jdbcUrl"
 			value="jdbc:log4jdbc:oracle:thin:@localhost:1521:XE"></property>
-		<property name="username" value="kmj"></property>
-		<property name="password" value="1234"></property>
+		<property name="username" value="scott"></property>
+		<property name="password" value="tiger"></property>
 	</bean>
 
 	<!-- HikariCP configuration -->
