+++ drinking01/src/main/webapp/WEB-INF/views/board/list.jsp
@@ -10,23 +10,30 @@
 <title>Insert title here</title>
 </head>
 <body>
+<form method = "get" action = "/board/register">
+
 <div>
 	<h2> List Page</h2>
 </div>
+<div>
+	<input type = "submit" value = "글쓰기">
+</div>
+
+</form>
 
 <div>
-	<table width = "100%">
+	<table>
 		<tr>
 			<th>글번호</th>
 			<th>글제목</th>
 			<th>작성자</th>
 			<th>작성일</th>
 		</tr>
-		<c:forEach var="item" items = "${boardList}" >
+		<c:forEach var="board" items = "${boardList}" >
 		<tr>
-			<td>${item.bno}</td>
-			<td>${item.title}</td>
-			<td>${item.writer}</td>
+			<td>${board.bno}		
+			<td><a href = "/board/get?bno=<c:out value = '${board.bno}'/>">${board.title }</a></td>
+			<td>${board.writer}</td>
 			<td><fmt:formatDate pattern="yyyy-MM-dd" value="${item.regDate}"/></td>
 		<tr>
 		</c:forEach>
