+++ drinking01/src/main/webapp/WEB-INF/web.xml
@@ -29,5 +29,29 @@
 		<servlet-name>appServlet</servlet-name>
 		<url-pattern>/</url-pattern>
 	</servlet-mapping>
+	
+	   <filter>
+      <filter-name>encodingFilter</filter-name>		 
+      <filter-class>org.springframework.web.filter.CharacterEncodingFilter
+      </filter-class>
+      <init-param>
+         <param-name>encoding</param-name>
+         <param-value>UTF-8</param-value>
+      </init-param>
+   </filter>
+   <filter-mapping>
+      <filter-name>encodingFilter</filter-name>
+      <url-pattern>/*</url-pattern>
+   </filter-mapping>
+   
+   <filter>
+   	<filter-name>springSecurityFilterChain</filter-name>
+   	<filter-class>org.springframework.web.filter.DelegatingFilterProxy</filter-class>
+   </filter>
+   
+   <filter-mapping>
+   	<filter-name>springSecurityFilterChain</filter-name>
+   	<url-pattern>/*</url-pattern>
+   </filter-mapping>
 
 </web-app>
