+++ drinking01/src/test/java/com/first/mapper/BoardMapperTest.java
@@ -30,8 +30,23 @@ public class BoardMapperTest {
 //	@Test
 //	public void insertDummies()
 //	{
-//		BoardVO board = new Boa
-//		mapper.insert();
+//		BoardVO board = new BoardVO();
+//		for( int i = 0; i < 100; i++)
+//		{
+//			board.setWriter("user"+ i);
+//			board.setTitle("Test Title" + i);
+//			board.setContent("Test Content" + i);
+//			mapper.insert(board);
+//		}
 //	}
-
+	
+	@Test
+	public void testGetBoard()
+	{
+		Long bno = 50L;
+		
+		log.info(mapper.getBoard(bno));
+		
+	}
+	
 }
